fix(gui): rule the per-layout override out, and guard the panel's wiring
Review of the lane found four majors, all of them things the
census and parity guards structurally cannot see.

The master read On over a glassless monocle App Bar. A
per-layout `liquid_glass` shadows the global, the master cannot
clear one, and reporting it would state a disagreement while
withholding the control that ends it — so the divergence
sentence is about the three GLOBAL leaves, ruled explicitly
rather than left to be rediscovered. That matches every other
per-layout override: the App Bar thickness slider says nothing
about `monocle.set_app_bar_thickness` either.

The panel's half had no guard at all. `LiquidGlassMasterTests`
never leaves the model and the view is AppKit-hosted, so
hardcoding either the controller's read or the view's hand-off
left the row writing a leaf nothing renders — the user-visible
half of #1307 gone, suite green. Two needles, because the read
and the hand-off fail separately.

`roundTrip` did not seed the new leaf, so deleting the decode
line passed while every stored `true` silently reverted on load;
the encoded sub-key and its default are pinned beside the
refusal group's. Seeding it took that file past §2.1's ceiling,
so the decode half is now its own suite — the shape pin and the
what-a-file-decodes-to pins were two subjects sharing a file.

`ColorsRowOrder.glassAtRest` was read by nothing — the card
hand-drew its row — so a second glass row would have rendered
nowhere and redded nothing. The card consumes the list now.

And `masterFansOut`'s `false` pass started where it meant to
end, so a setter writing nothing passed it. It seeds the
opposite first.
